[#30]feature: API Instance 수정
