+++ src/app/page.tsx
@@ -1,4 +1,3 @@
-"use client";
 import Image from "next/image";
 
 export default function Home() {
